口座削除に参照判定追加

+++ app/controllers/api/v1/user/bank_accounts_controller.rb
@@ -62,6 +62,10 @@ class Api::V1::User::BankAccountsController < ApplicationController
 
   def destroy
     account_id = params[:id].to_i
+    if disallowed_at_bank_ids?([account_id])
+      render_disallowed_financier_ids && return
+    end
+
     if @current_user.try(:at_user).try(:at_user_bank_accounts).pluck(:id).include?(account_id)
       Services::AtUserService.new(@current_user).delete_account(Entities::AtUserBankAccount, [account_id])
     end
